--- src/core/README.md
@@ -1,7 +1,0 @@
-This core part of the project attempts compliance with the [Yarn 2.0 language specification](https://github.com/YarnSpinnerTool/YarnSpinner/blob/9275277f50a6acbe8438b29596acc8527cf5581a/Documentation/Yarn-Spec.md).
-
-Its API is similar to the bondage.js for Yarn 1.x API.
-
-Non-spec features (like markup) should be implemented in a separate layer.
-
-In theory, you could run this alone and have a "vanilla" Yarn 2.0 experience.
