--- src/core/index.js
@@ -1,8 +1,0 @@
-import runner from './runner';
-import results from './results';
-
-runner.OptionsResult = results.OptionsResult;
-runner.TextResult = results.TextResult;
-runner.CommandResult = results.CommandResult;
-
-export default runner;
