+++ src/default-variable-storage.js
