+++ src/lexer/lexer-state.js
