+++ src/lexer/states.js
