+++ src/parser/compiled-parser.js
